feat!: moved database into its own monorepo lib package
Needed to add https://www.npmjs.com/package/tsconfig-paths to backend to allow ts-node to parse the @database tsconfig path. Shame that this doesn't work out of the box.
