+++ apps/backend/.env.example
@@ -1,2 +1,1 @@
 PORT=3001
-DATABASE_URL="postgresql://postgres:postgres@localhost:5555/my-db?schema=public"
