+++ apps/backend/package.json
@@ -5,10 +5,8 @@
     "main": "index.js",
     "scripts": {
         "check": "tsc",
-        "cold-start": "npm run migrate && npm run seed && npm run start",
-        "migrate": "npx prisma migrate dev --name init",
-        "start": "ts-node src/index.ts",
-        "seed": "ts-node src/tmpReadModels.ts"
+        "start": "ts-node -r tsconfig-paths/register src/index.ts",
+        "seed": "ts-node -r tsconfig-paths/register src/tmpReadModels.ts"
     },
     "keywords": [],
     "author": "",
@@ -27,8 +25,8 @@
         "@types/express": "^5.0.1",
         "@types/morgan": "^1.9.9",
         "@types/node": "^22.14.1",
-        "prisma": "^6.6.0",
         "ts-node": "^10.9.2",
+        "tsconfig-paths": "^4.2.0",
         "typescript": "^5.8.3"
     }
 }
