+++ libs/database/.env.example
@@ -1,0 +1,1 @@
+DATABASE_URL="postgresql://postgres:postgres@localhost:5555/my-db?schema=public"
