+++ tsconfig.json
@@ -1,11 +1,15 @@
 {
     "compilerOptions": {
+        "baseUrl": ".",
         "esModuleInterop": true,
         "forceConsistentCasingInFileNames": true,
         "strict": true,
         "skipLibCheck": true,
-        "noEmit": true
+        "noEmit": true,
+        "paths": {
+            "@database/*": ["libs/database/src/*"]
+        }
     },
     "exclude": ["node_modules"],
-    "references": [{ "path": "apps/frontend" }, { "path": "apps/backend" }]
+    "references": [{ "path": "apps/frontend" }, { "path": "apps/backend" }, { "path": "libs/database" }]
 }
